+++ PyCheMPS2/tests/test14.py
@@ -34,45 +34,41 @@ Ham = PyCheMPS2.PyHamiltonian( -1, psi4group, orbirreps, filename )
 
 # Define the symmetry sector
 TwoS = 0     # Two times the targeted spin
-N = 14       # The number of electrons
+Nelec = 14   # The number of electrons
 Irrep = 0    # The targeted irrep
 
 # Define the CASSCF
 DOCC  = np.array([ 3, 0, 0, 0, 0, 2, 1, 1 ], dtype=ctypes.c_int) # see N2.ccpvdz.out
 SOCC  = np.array([ 0, 0, 0, 0, 0, 0, 0, 0 ], dtype=ctypes.c_int)
 NOCC  = np.array([ 1, 0, 0, 0, 0, 1, 0, 0 ], dtype=ctypes.c_int)
-NDMRG = np.array([ 4, 0, 1, 1, 0, 4, 1, 1 ], dtype=ctypes.c_int)
-NVIRT = np.array([ 2, 1, 2, 2, 1, 2, 2, 2 ], dtype=ctypes.c_int)
+NDMRG = np.array([ 2, 0, 1, 1, 0, 2, 1, 1 ], dtype=ctypes.c_int)
+NVIRT = np.array([ 4, 1, 2, 2, 1, 4, 2, 2 ], dtype=ctypes.c_int)
 theDMRGSCF = PyCheMPS2.PyCASSCF(Ham, DOCC, SOCC, NOCC, NDMRG, NVIRT)
 
-# Setting up the ConvergenceScheme
-# setInstruction(instruction, D, Econst, maxSweeps, noisePrefactor)
-OptScheme = PyCheMPS2.PyConvergenceScheme(1) # 1 instruction
-OptScheme.setInstruction(0, 1000, 1e-8, 20, 0.0)
-
 # Setting the DMRGSCFoptions and run DMRGSCF
-rootNum = 1 # Ground state only
-theDMRGSCFoptions = PyCheMPS2.PyDMRGSCFoptions()
-theDMRGSCFoptions.setDoDIIS(True)
-theDMRGSCFoptions.setWhichActiveSpace(2) # 2 means localized orbitals
-Energy = theDMRGSCF.solve(N, TwoS, Irrep, OptScheme, rootNum, theDMRGSCFoptions)
+root_num = 1 # Ground state only
+scf_options = PyCheMPS2.PyDMRGSCFoptions()
+scf_options.setDoDIIS( True )
+IPEA = 0.0
+IMAG = 0.0
+Energy1 = theDMRGSCF.solve_fci( Nelec, TwoS, Irrep, root_num, scf_options)
+Energy2 = theDMRGSCF.caspt2_fci(Nelec, TwoS, Irrep, root_num, scf_options, IPEA, IMAG)
 
 # Clean-up
-if theDMRGSCFoptions.getStoreUnitary():
+if scf_options.getStoreUnitary():
     theDMRGSCF.deleteStoredUnitary()
-if theDMRGSCFoptions.getStoreDIIS():
+if scf_options.getStoreDIIS():
     theDMRGSCF.deleteStoredDIIS()
 
 # The order of deallocation matters!
-del theDMRGSCFoptions
-del OptScheme
+del scf_options
 del theDMRGSCF
 del Ham
 del Initializer
 
 # Check whether the test succeeded
-if (np.fabs(Energy + 109.15104350802) < 1e-8):
-    print "================> Did test 9 succeed : yes"
+if (( np.fabs( Energy1 + 109.103502335253 ) < 1e-8 ) and ( np.fabs( Energy2 + 0.159997813112638 ) < 1e-8 )):
+    print "================> Did test 14 succeed : yes"
 else:
-    print "================> Did test 9 succeed : no"
+    print "================> Did test 14 succeed : no"