+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 * Class TensorOperator
 * Class ThreeDM: DMRG 3-RDM in O(L^4 D^3 + L^6 D^2)
 * Class Cumulant: Cumulant reconstruction of the DMRG 4-RDM
-* Class CASPT2: Routines for the CASPT2-N energy (IPEA follows)
+* Class CASPT2: Internally contracted CASPT2
 * DMRG diagonal 4-RDM in DMRG::Diag4RDM
 * Sweeps after RDM calculation allowed
 * FCI 3-RDM in FCI::Fill3RDM